Add guard clause for missing files
Currently the runner tries to inspect files even if they do not exist.
This diff prevents that.

+++ lib/pronto/runner.rb
@@ -53,6 +53,7 @@ module Pronto
     end
 
     def ruby_executable?(path)
+      return false unless File.exist?(path)
       return false if File.directory?(path)
       line = File.open(path, &:readline)
       line =~ /#!.*ruby/
